Move derive_nullifier from gadget.rs to derive_nullifier.rs
