+++ src/circuit.rs
@@ -50,6 +50,7 @@ mod circuit_vanilla;
 mod circuit_zsa;
 
 pub(in crate::circuit) mod commit_ivk;
+pub(in crate::circuit) mod derive_nullifier;
 pub(in crate::circuit) mod gadget;
 pub(in crate::circuit) mod note_commit;
 pub(in crate::circuit) mod orchard_sinsemilla_chip;
